fixed linux and a bunch of warnings
Former-commit-id: e47c82301164beafa0d5dc7c1007c03e81d7a1e6

+++ src/shader_loader.cpp
@@ -367,7 +367,7 @@ namespace rendering
             }
             else
             {
-                uniform.array_size = strtol(*rest, rest, 10);
+                uniform.array_size = (i32)strtol(*rest, rest, 10);
             }
         }
     }
@@ -706,7 +706,7 @@ namespace rendering
                         if (rest[0] == '=')
                         {
                             rest++;
-                            vertex_attribute->location = strtol(rest, &rest, 10);
+                            vertex_attribute->location = (i32)strtol(rest, &rest, 10);
 
                             while (rest[0] == ' ' || rest[0] == ')')
                                 rest++;
